refactor: route dom-core imports through dom-api barrel (51 -> 5 fan-in)
Closes #633
