+++ src/utils/flow-view-rendering.js
@@ -12,7 +12,7 @@ import {
 } from './flow-view-helpers.js';
 import { createCategoryGroup } from './flow-category-renderer.js';
 import { createFlowCard } from './flow-card-setup.js';
-import { _el } from './dom-core.js';
+import { _el } from './dom-api.js';
 import { buildDomainButtonBar } from './dom-buttons.js';
 import { buildViewHeader } from './view-header.js';
